requirements.txt
# Production requirements
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
mysqlclient==2.2.4
pandas==2.2.3
gunicorn==22.0.0
python-dotenv==1.0.1

# Development/Optional
whitenoise==6.7.0